add a fluent builder for scope state
Maybe silly, but writing all this feels like a lot of work:

    var state = new Dictionary<string, object?>();
    state["order_id"] = order.Id;
    state["order_status"] = order.Status;
    using var _ = _logger.BeginScope(state);
    _logger.LogInformation("Accepted order.");

This is nicer:

    _logger.LogInformation("Accepted order {order_id}.", order.Id);

But it relies on putting every single property in your actual log
message. It often doesn't make sense to do this. And, it's too easy to
screw up the order of the template arguments; especially when editing
log messages later.

So now:

    _logger.With("order_id", order.Id).LogInformation("Accepted order");

Or:

    using var _ = _logger
        .With("order_id", order.Id)
        .With("order_status", order.Status)
        .BeginScope();
    _logger.LogInformation("Accepted order.");